refactor(components): 重构任务摘要功能
- 创建新的 TaskSummaryManager 类来管理任务摘要生成和复制功能
- 从 CalendarView 类中移除与任务摘要相关的代码
- 优化了任务摘要的生成和复制逻辑
- 新增了富文本、Markdown 和纯文本三种复制格式
